generation: Allow test cases to specify a destination folder
If we want to generate boilerplate for public APIs such as a `name.js` and
`length.js` test for every built-in function, we'd need to have templates
like `name.template` and `length.template`, and test case files for every
built-in function, such as `Array.prototype.at.case`.

This poses a problem, because we'd want the generated tests to be like
`test/built-ins/Array/prototype/at/name.js`. Currently, the test
generation tool only does it the other way around, with the test template
supplying the prefix and the test case supplying the name: e.g.,
`name/Array.prototype.at.js`.

This allows a `path` key in test case frontmatter which, if present,
generates the test in that folder, while naming the generated test after
the template.

The `path` key in the test template frontmatter is still used as a prefix
if present, but now it is allowed to be absent.